requirements.txt
altair
matplotlib
numpy
pandas
PyYAML
scikit-learn
scipy
seaborn
streamlit
